requirements.txt
flask==2.2.2
openai==0.27.4
redis==3.5.3
redis-py-cluster==2.1.3
tiktoken==0.3.3
urllib3==1.24.3
werobot==1.13.1
beautifulsoup4==4.9.3
aiohttp==3.8.4
baidu-aip==4.16.10
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bottle==0.12.25
ca-certificates==2023.01.10
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.0.4
et==0.0.2
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.0.1
jinja2==3.1.2
libcxx==14.0.6
libffi==3.4.2
markupsafe==2.1.1
multidict==6.0.4
ncurses==6.4
openssl==1.1.1t
pip==23.0.1
python==3.9.16
readline==8.2
regex==2023.3.23
requests==2.28.2
setuptools==65.6.3
sqlite==3.41.2
tk==8.6.12
tqdm==4.65.0
tzdata==2023c
werkzeug==2.2.3
wheel==0.38.4
xmltodict==0.13.0
xz==5.2.10
yarl==1.8.2
zipp==3.11.0
zlib==1.2.13
